check working dir is dirty in bash

+++ deploy.sh
@@ -2,6 +2,11 @@
 
 echo -e "\033[0;32mDeploying updates to GitHub...\033[0m"
 
+if [ "$(git status -s)" ]; then
+    echo "The working directory is dirty. Please commit any pending changes."
+    exit 1;
+fi
+
 # Remove previous published contents.
 rm -rf public
 mkdir public
